+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+ - Added queues for backup shipping
+
 ### Fixed
 
 - Fixed UID-mismatch when creating rollback entries in K8s/CRD